+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - adds `indicatorGradient` to `ToggleStyle` ([#44](https://github.com/splashbyte/animated_toggle_switch/issues/44))
 - fixes problems with `backgroundGradient` and `backgroundColor` ([#46](https://github.com/splashbyte/animated_toggle_switch/issues/46))
+- adds `AnimationType.none`
+- adds `animateThemeChanges` to `AnimatedToggleSwitch`
 
 ## 0.8.0 (2023-09-02)